requirements.txt
# ===============================================
# METEO TOULOUSE PLATFORM - REQUIREMENTS
# ===============================================

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Web Interface
streamlit>=1.28.0
plotly>=5.15.0
dash>=2.14.0

# APIs et Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.8.0

# Kafka
kafka-python>=2.0.2
confluent-kafka>=2.2.0

# Spark
pyspark>=3.4.0
findspark>=2.0.1

# Bases de données
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.7
pymongo>=4.5.0
sqlite3  # Built-in

# Machine Learning / LLM
transformers>=4.33.0
torch>=2.0.0
scikit-learn>=1.3.0
huggingface-hub>=0.16.0

# NLP
nltk>=3.8.0
spacy>=3.6.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1
configparser>=5.3.0

# Logging et Monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Date/Time
python-dateutil>=2.8.2
pytz>=2023.3

# Utilities
tqdm>=4.66.0
click>=8.1.0
pathlib2>=2.3.7

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter (pour exploration)
jupyter>=1.0.0
ipykernel>=6.25.0

# Visualization avancée
seaborn>=0.12.0
matplotlib>=3.7.0

# APIs spécialisées météo
pyowm>=3.3.0  # OpenWeatherMap

# Excel/CSV avancé
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Scheduling
schedule>=1.2.0
celery>=5.3.0
redis>=4.6.0

# HTTP server
fastapi>=0.103.0
uvicorn>=0.23.0

# Data validation
pydantic>=2.3.0
marshmallow>=3.20.0